API

API

CorrelationExpansion.state.State — Type.

State(operators, correlations[, factor])
State(operators, masks)
State(basis, masks)
State(basis_l, basis_r, masks)

State representing a correlation expansion series.

Physically the state consists of a product state $ρ₁$ and the selected correlations $σˢ$. Which correlations are included is specified by the given masks s ∈ Sₙ.

\[ρ = ρ₁ + ∑_{s ∈ Sₙ} \tilde{σ}ˢ\]

source

CorrelationExpansion.state.correlation — Function.

correlation(rho, mask; <keyword arguments>)

Calculate the correlation of the subsystems specified by the given mask.

Arguments

rho: Density operator of the total system.

mask: Indices or mask specifying on which subsystems the given correlation is defined.

operators (optional): A tuple containing all reduced density operators of the single subsystems.

subcorrelations: A (mask->operator) dictionary storing already calculated correlations.

source

CorrelationExpansion.state.embedcorrelation — Function.

embedcorrelation(ops, mask, correlation)

Tensor product of a correlation and the density operators of the remaining subsystems.

Arguments

operators: Vector containing the reduced density operators of each subsystem.

mask: Indices or mask specifying on which subsystems the given correlation is defined.

correlation: Correlation operator defined in the specified subsystems.

source

CorrelationExpansion.state.approximate — Function.

approximate(rho[, masks])

Correlation expansion of a density operator.

If masks are specified, only these correlations are included.

source